Fix leading whitespace in tokens
Previous to this change, the optional spaces in number regexes causes
leading whitespace to be added to some parsed tokens.

The current tests are insufficient to catch this issue since they do not
include the parsed text of the nodes.

Given the following sample code:

    function int f(a);
        a = 0;
    endfunction

* The variable 'a' parsed as "\n    a" instead of "a"
* The unsigned number '0' parsed as " 0" instead of "0"
* The keyword 'endfunction' parsed as "\nendfunction" instead of "endfunction"

I am not sure why this happens but I suspect the optional whitespace in
the regexes has a negative interaction with the 'extras' field of the
grammar.

This change fixes the problem by breaking up the number regexes into a
sequence of regexes splitting on the optional whitespace.